requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.1
pandas==2.2.2
numpy==1.26.4
google-generativeai==0.7.2
openai==1.43.0
scikit-learn==1.5.2
llama-index
llama-index-core
llama-index-vector-stores-chroma
llama-index-embeddings-huggingface
llama-index-llms-google-genai
chromadb
torch
transformers
sentence-transformers

jinja2

